requirements.txt
Django==1.10.3
django-autoslug==1.9.3
django-bootstrap3==7.1.0
django-crispy-forms==1.6.1
Pillow==3.3.0
wheel==0.24.0
django-markitup==3.0.0
Markdown==2.6.7

